+++ src/main/resources/lessons/chromedevtools/documentation/ChromeDevTools_Assignment.adoc
