+++ src/main/resources/lessons/chromedevtools/documentation/ChromeDevTools_Assignment_Network.adoc
