+++ src/main/resources/lessons/chromedevtools/documentation/ChromeDevTools_console.adoc
