+++ src/main/resources/lessons/chromedevtools/i18n/WebGoatLabels.properties
