+++ src/main/resources/lessons/chromedevtools/images/ChromeDev_Console_Clear.jpg
